feat : 결제 정보 컴포넌트에서 컴포넌트 분리 및 약관 동의 시 버튼 활성화 로직 추가
