+++ nonsoolmateClient/src/pages/payment/types/paymentInfoType.ts
@@ -1,0 +1,16 @@
+export interface Discount {
+  id: number;
+  title: string;
+  rate: number;
+}
+
+export interface Plan {
+  id: number;
+  title: string;
+  price: number;
+  defaultDiscount: Discount[];
+}
+
+export interface OrderProps {
+  plan: Plan;
+}
